+++ src/hosted_db_runner/build_nsjail.sh
@@ -1,0 +1,9 @@
+#!/usr/bin/env bash
+
+git clone https://github.com/google/nsjail.git nsjail-checkout
+cd nsjail-checkout
+make
+mv nsjail ..
+cd ..
+rm -rf nsjail-checkout
+
